-- Fixing DrawLineSegment.

+++ core/src/main/java/leikr/Engine.java
@@ -396,7 +396,7 @@ public abstract class Engine extends BaseGamePadListener implements InputProcess
     }
 
     public final void drawLineSegment(int c, BigDecimal x1, BigDecimal y1, BigDecimal x2, BigDecimal y2) {
-        lScreen.drawLineSegment(c, x1.intValue(), y1.intValue(), x2.intValue(), y2.intValue());
+        lScreen.drawLineSegment(c, x1, y1, x2, y2);
     }
     //end shape drawing methods
